requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
customtkinter==5.1.2
darkdetect==0.8.0
datasets==2.12.0
deep-translator==1.10.1
dill==0.3.6
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
happytransformer==2.4.1
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.22.3
pyarrow==11.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
responses==0.18.0
sentencepiece==0.1.98
six==1.16.0
soupsieve==2.4.1
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
xxhash==3.2.0
yarl==1.9.2
